static cast

+++ engine/src/flutter/impeller/runtime_stage/runtime_stage.cc
@@ -88,8 +88,8 @@ absl::StatusOr<RuntimeStage> RuntimeStage::Create(
       }
       if (i->struct_elements()) {
         for (const auto& elem : *i->struct_elements()) {
-          desc.struct_elements.emplace_back(
-              StructElement{elem->name()->str(), elem->byte_size()});
+          desc.struct_elements.emplace_back(StructElement{
+              elem->name()->str(), static_cast<size_t>(elem->byte_size())});
         }
       }
       desc.struct_float_count = i->struct_float_count();
